+++ libleptongui/include/integer_combo_box.h
@@ -1,7 +1,7 @@
 /* Lepton EDA Schematic Capture
  * Copyright (C) 1998-2010 Ales Hvezda
  * Copyright (C) 1998-2013 gEDA Contributors
- * Copyright (C) 2017-2024 Lepton EDA Contributors
+ * Copyright (C) 2017-2025 Lepton EDA Contributors
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -52,6 +52,8 @@ struct _SchematicIntegerComboBox
   gboolean changed;
 };
 
+G_BEGIN_DECLS
+
 GtkEntry*
 schematic_integer_combo_box_get_entry (GtkWidget *widget);
 
@@ -70,3 +72,4 @@ schematic_integer_combo_box_set_model (GtkWidget *widget,
 void
 schematic_integer_combo_box_set_value (GtkWidget *widget,
                                        int value);
+G_END_DECLS