+++ libleptongui/src/o_text.c
@@ -1,7 +1,7 @@
 /* Lepton EDA Schematic Capture
  * Copyright (C) 1998-2010 Ales Hvezda
  * Copyright (C) 1998-2016 gEDA Contributors
- * Copyright (C) 2017-2024 Lepton EDA Contributors
+ * Copyright (C) 2017-2025 Lepton EDA Contributors
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -30,56 +30,6 @@
 #endif
 
 #include "schematic.h"
-
-/*! \todo Finish function documentation!!!
- *  \brief
- *  \par Function Description
- *
- */
-void
-o_text_prepare_place (SchematicWindow *w_current,
-                      char *text,
-                      int color,
-                      int align,
-                      int rotate,
-                      int size)
-{
-  SchematicCanvas *page_view = schematic_window_get_current_canvas (w_current);
-  g_return_if_fail (page_view != NULL);
-
-  LeptonPage *page = schematic_canvas_get_page (page_view);
-  if (page == NULL) {
-    return;
-  }
-
-  /* Insert the new object into the buffer at world coordinates (0,0).
-   * It will be translated to the mouse coordinates during placement. */
-
-  w_current->first_wx = 0;
-  w_current->first_wy = 0;
-
-  /* remove the old place list if it exists */
-  lepton_object_list_delete (page->place_list);
-  page->place_list = NULL;
-
-  /* here you need to add OBJ_TEXT when it's done */
-  page->place_list =
-    g_list_append(page->place_list,
-                  lepton_text_object_new (color,
-                                          0,
-                                          0,
-                                          align,
-                                          rotate, /* zero is angle */
-                                          text,
-                                          size,
-                                          /* has to be visible so you can place it */
-                                          /* visibility is set when you create the object */
-                                          VISIBLE,
-                                          SHOW_NAME_VALUE));
-
-  i_action_start (w_current);
-  i_set_state (w_current, TEXTMODE);
-}
 
 
 /*! \todo Finish function documentation!!!